osx friendly tests
The ManagedProcessTest relies on the executables that are commonly found on different
OSes.  Although many of the BSD and Linux executables are similar, they have some minor
differences.  For example, ls -4 returns with an exit code of 2 on OSX, not 1 like on Linux.
Similarly, sleep just takes a number of seconds on OSX, unlike on Linux where it also takes
a unit (eg 10s or 10ms).